requirements.txt
django-import-export==2.4.0
about==5.2
appdirs==1.4.4
asgiref==3.2.10
auth==0.5.3
beautifulsoup4==4.9.1
blinker==1.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.9
django-admin-interface==0.13.2
django-colorfield==0.3.2
django-crispy-forms==1.9.2
django-etc==1.2.0
django-filter==2.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-hitcount==1.3.3
django-like-button==0.0.7
django-likes==2.0.1
django-rest-auth==0.9.2
django-secretballot==1.0.0
djangorestframework==3.11.1
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.0.1
eventlet==0.26.1
falcon==2.0.0
filelock==3.0.12
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
Jinja2==3.0.0a1
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
mongoengine==0.20.0
monotonic==1.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
passlib==1.7.2
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.6.1
PyJWT==1.7.1
pymdown-extensions==7.1
pymongo==3.11.0
python-social-auth==0.2.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
register==0.1
requests==2.24.0
requests-oauthlib==0.3.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
virtualenv==20.0.28
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0